Remove MissingImageIdTestTrait
